requirements.txt
streamlit>=1.31.0
snowflake-connector-python>=3.7.0
azure-storage-blob>=12.19.0
Pillow>=10.2.0
pandas>=2.2.0
